Added dummy function to list locations
I first defined the character vector containing location names in the _targets.R script, but this resulted in wrong encoding and messed up Norwegian letters (even when setting the encoding).  Turns out it works just fine as long as the vector is defined and encoding set WITHIN a function that is part of the targets workflow.