Corrected messed up area names in FeFo data
verbatimLocality is set to "Fefo" for all transects in the "FeFo" data. This seems to be an issue with the current version of the data and should be fixed in the database. 
For in the meantime,  I am retrieving the correct area names from the column locationRemarks, and also fixing "Stalluvarre" as being a locality within "Indre Finnmark" instead of an area.