Adjusted prior for initial density
Changed from log scale to natural scale, with a lower bound of 0. This was the easiest way to deal with the fact that after removing the scaling parameters, density estimates became A LOT smaller (as they are now per m^2 instead of km^2).  This fixed the issue with initial density getting estimated at lower prior bound.