Wrote function to plot distribution of yearly reproduction
This visualizes why the mixing of the RE model for R is so abyssmal: the distribution of the annual R under the current assumptions seems to be bimodal. 
We will try to assemble data on rodent abundance and see if accounting for low vs. high rodent years can help with this.